# lib/api-zod/src/generated/types/botStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotStatus {
  isRunning: boolean;
  /** @nullable */
  currentAccount: string | null;
  /** @nullable */
  lastRunAt: string | null;
  totalSearchesToday: number;
  /** @nullable */
  activeRunId?: string | null;
}