# lib/api-zod/src/generated/types/deskAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeskAccountStatus } from "./deskAccountStatus";

export interface DeskAccount {
  id: string;
  email: string;
  name: string;
  status: DeskAccountStatus;
  totalPoints: number;
  todayPoints: number;
  /** @nullable */
  lastRun: string | null;
  searchesCompleted?: number;
}