# lib/api-zod/src/generated/types/runInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RunInput {
  /** Specific account IDs to run (omit to run all) */
  accountIds?: string[];
}