# lib/api-zod/src/generated/types/runLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunLogStatus } from "./runLogStatus";

export interface RunLog {
  id: string;
  accountId: string;
  accountName: string;
  timestamp: string;
  searchesDone: number;
  pointsEarned: number;
  status: RunLogStatus;
  /** @nullable */
  errorMessage?: string | null;
}